requirements.txt
numpy==1.26.4
pyserial==3.5
websockets==12.0
pytest==8.4.2
pyserial==3.5